Removed unnecessary locking, fixed flow expiration and memory freeing bugs.
The locking discipline has recently changed in OpenFlow which now synchronizes on control messages, such that modifications to the table can only occur one at a time. Further, datapath packets arrive already wrapped in a rcu_read_lock block, thus the only locking that needs to be done in the hardware table is getting the mutex during a timeout operation, because it is not called from the control path locking section.

A bug was fixed that was causing flows to timeout before they should, thanks to Sivasankar Radhakrishnan for this code. Also a bug was fixed that was causing the flow's private data structure memory to be freed rather than returned to the free list.