Spec: Define behavior for conflicting flow mods
Define a CHECK_OVERLAP flag for flow mods which requires the switch to
check for conflicting entries with the same priority, i.e., those for
which at least one packet could match multiple entries.

This flags is intended to be used for debugging purposes; in the worst
case, it may require a linear check through the flow table.