Bugfix : report error when add-queue for non-existing port.
This caused segfault instead of reporting the error.
The problem was that dp_lookup_port doesn't provide any sense
whether a port exists in the datapath. It refers to the
DP_MAX_PORTS entries to the switch, not to currently
configured ports in the switch. Instead of checking
p we check p->netdev and detect non-queue case.